5th. StatisticsGeneratorNew: correct rounding method to get/out type avgExamScore = 4.32 (f.e.).

+++ src/main/java/org/example/utils/StatisticsGeneratorNew.java
@@ -54,7 +54,8 @@ public class StatisticsGeneratorNew {
                     .average();
             statistics.setAvgExamScore(0);
             avgExamScore.ifPresent(value -> statistics.setAvgExamScore(
-                    (float)BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP).doubleValue()));
+//                    (float)
+                            BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP).doubleValue()));
 //                    (float) BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP).doubleValue()));
         });
